Create GradetoLetterTranslator
This is my first project from the list found here: https://www.reddit.com/r/Python/comments/9m5wpk/50_programming_problems_for_an_introductory/

The goal is to take a grade input and output it to a Ryerson letter grade.  I am making the problem a bit more complicated by using BeautifulSoup to scrape the list data and a few extra checks to make certain that the input data is both valid and in range.  The final function matches the letter grade against a lower_bound grade number list and returns the matching letter grade with the appropriate grade modifier.  This could easily be done with an array and a bunch of nested if statements, but I decided to forgo using anything like a CASE statement in favor of manipulating values and then using list references.