Create NumberGame
Clunky - one of my first projects in Python.  This is meant to be a number guessing game with feedback to the user to help make the process of guessing a bit easier.  It is super buggy, though.  I need to go back through it once I have a better handle on the Python in general.