Create find-matching-lists
Given a list of elements, create and return a list whose elements are lists that contain the consecutive runs of equal
elements of the original list. Note that elements that are not duplicated in the original list should become singleton
lists in the result, so that every element gets included in the resulting list of lists.