Create Rec-Sequence
Given a number of elements, n, calculate the rec sequence from a1 = 1 through to the end of the sequence a.n.  Following
the rules of the rec sequence, the lower value is only added if it is > 0 and not in the sequence already, otherwise
the higher value is added.  To make life easier (and because I am only interested in the math this time) all n values
will be greater than 1.